Fix inheritance for `intro` section
It looks like there are restrictions on where a template section variable can be defined, that I'm not fully certain I understand. However, previously with the intro section defined where it was, the intro was not rendered unless it was specified in an implementing template (i.e. not falling back to translated intro). By moving the defaulting of the intro section up a level in the template hierarchy it is rendered correctly in all cases.